+++ scripts/modtools/equip-item.lua
@@ -5,7 +5,7 @@
 modtools/equip-item
 ===================
 Force a unit to equip an item with a particular body part; useful in
-conjunction with the ``create`` scripts above.  See also `forceequip`.
+conjunction with the ``create-*`` scripts.  Replaces ``forceequip``.
 
 =end]]
 local utils = require 'utils'