+++ NEWS.rst
@@ -178,6 +178,13 @@ Misc Improvements
 Removed
 -------
 - `embark-tools` nano: 1x1 embarks are now possible in vanilla 0.40.24
+- dwarfexport: has been disabled since 40.01
+- mapexport: has been disabled since 40.01
+- createitem: replaced by `modtools/create-item`
+- forceequip: replaced by `modtools/equip-item`
+- misery: not compatible with DF 40.01 or later
+- df2mc: it's been broken for years
+
 
 DFHack 0.40.24-r3
 =================